paging implemented

+++ app/src/main/java/com/picobyte/flantern/types/Message.kt
@@ -3,7 +3,7 @@ package com.picobyte.flantern.types
 data class Message(
     val user: String? = null,
     val content: String? = null,
-    val timestamp: Int? = null,
+    val timestamp: Long? = null,
     val replying: String? = null,
     val forwarded: Boolean? = null
 )
